# Increase BGZip streaming decompression performance