… tests/data/test.fastq.bgzip.gz