### Checklist
- [x] Pull request details were added to CHANGELOG.rst
- [ ] Documentation was updated (if needed)

Before
```
Benchmark 1: python -m isal.igzip -cd ~/test/5millionreads_R1.fastq.gz > /dev/null
  Time (mean ± σ):      2.201 s ±  0.017 s    [User: 2.145 s, System: 0.055 s]
  Range (min … max):    2.184 s …  2.228 s    5 runs

Benchmark 1: python -m isal.igzip -cd ~/test/5millionreads_R1.fastq.bgzip.gz > /dev/null
  Time (mean ± σ):      2.633 s ±  0.018 s    [User: 2.565 s, System: 0.067 s]
  Range (min … max):    2.608 s …  2.655 s    5 runs
```
This is the same file. Gzip level 1 and bgzip level 1. Overhead for decompressing the bgzip file 400 ms.

After:
```
Benchmark 1: python -m isal.igzip -cd ~/test/5millionreads_R1.fastq.bgzip.gz > /dev/null
  Time (mean ± σ):      2.340 s ±  0.012 s    [User: 2.269 s, System: 0.071 s]
  Range (min … max):    2.324 s …  2.354 s    5 runs
```
The overhead is reduced to 140 ms. This is acceptable since a header and a trailer needs to be decompressed every 16K or so and this happens in pure python code rather than C code.

A further reduction could probably be achieved by using the bgzip block length as stored in the header, but this is massively more complicated and the speed gain is going to be less than 10% at maximum given the decompression baseline for the normal gzip file. Also it is out of scope for this project.

@marcelm, FYI. Building a dedicated bgzip decompressor is probably not worth it for xopen.